swagger simplified
