+++ libs/server/mod.ts
@@ -4,3 +4,5 @@ export * from "./middlewares/cors.ts";
 export * from "./router/router.ts";
 export * from "./router/rest-handler.ts";
 export * from "./router/request-handler.ts";
+export * from "./middlewares/swagger.ts";
+export * from "./middlewares/scalar.ts";
